+++ NEWS.md
@@ -1,5 +1,9 @@
 # simmer 3.6.4.9000
 
+## New features:
+
+* `set_attribute()` (and `set_global()` by extension) can set multiple attributes at once by providing vectors of `keys` and `values` (or functions returning such `keys` and/or `values`). `get_attribute()` (and `get_global()` by extension) can retrieve multiple `keys` (#122).
+
 ## Minor changes and fixes:
 
 * Restore `ostream` after formatting (9ff11f8).